backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pyyaml
pytest
